+++ projects/packages/jetpack-mu-wpcom/src/features/wpcom-admin-bar/wpcom-admin-bar.php
@@ -7,10 +7,29 @@
  * @package automattic/jetpack-mu-wpcom
  */
 
+use Automattic\Jetpack\Connection\Manager as Connection_Manager;
 use Automattic\Jetpack\Jetpack_Mu_Wpcom;
 
-if ( get_option( 'wpcom_admin_interface' ) !== 'wp-admin' ) {
-	return;
+define( 'WPCOM_ADMIN_BAR_UNIFICATION', true );
+
+/**
+ * Adds the origin_site_id query parameter to a URL.
+ *
+ * @param string $url The URL to add the query param to.
+ * @return string The URL with the origin_site_id query parameter mey be added.
+ */
+function maybe_add_origin_site_id_to_url( $url ) {
+	$site_id = Connection_Manager::get_site_id();
+	if ( is_wp_error( $site_id ) ) {
+		return $url;
+	}
+
+	// Add query param to URL only for users who can access wp-admin.
+	if ( ! is_user_member_of_blog() ) {
+		return $url;
+	}
+
+	return add_query_arg( 'origin_site_id', $site_id, $url );
 }
 
 /**
@@ -41,40 +60,44 @@ function wpcom_enqueue_admin_bar_assets() {
 				#wpadminbar .quicklinks #wp-admin-bar-top-secondary #wp-admin-bar-search {
 					order: -1;
 				}
+
+				#wpadminbar .quicklinks #wp-admin-bar-top-secondary #wp-admin-bar-help-center {
+					order: 1;
+				}
 CSS
 		);
 	}
 }
-add_action( 'admin_bar_menu', 'wpcom_enqueue_admin_bar_assets' );
+add_action( 'wp_enqueue_scripts', 'wpcom_enqueue_admin_bar_assets' );
+add_action( 'admin_enqueue_scripts', 'wpcom_enqueue_admin_bar_assets' );
 
 /**
- * Repurposes the WP logo as a link to /sites.
+ * Replaces the WP logo as a link to /sites.
  *
  * @param WP_Admin_Bar $wp_admin_bar The WP_Admin_Bar core object.
  */
-function wpcom_repurpose_wp_logo_as_all_sites_menu( $wp_admin_bar ) {
+function wpcom_replace_wp_logo_with_wpcom_all_sites_menu( $wp_admin_bar ) {
 	foreach ( $wp_admin_bar->get_nodes() as $node ) {
 		if ( $node->parent === 'wp-logo' || $node->parent === 'wp-logo-external' ) {
 			$wp_admin_bar->remove_node( $node->id );
 		}
 	}
-
 	$wp_admin_bar->remove_node( 'wp-logo' );
 	$wp_admin_bar->add_node(
 		array(
-			'id'    => 'wp-logo',
+			'id'    => 'wpcom-logo',
 			'title' => '<span class="ab-icon" aria-hidden="true"></span><span class="screen-reader-text">' .
 						/* translators: Hidden accessibility text. */
 						__( 'All Sites', 'jetpack-mu-wpcom' ) .
 						'</span>',
-			'href'  => 'https://wordpress.com/sites',
+			'href'  => maybe_add_origin_site_id_to_url( 'https://wordpress.com/sites' ),
 			'meta'  => array(
 				'menu_title' => __( 'All Sites', 'jetpack-mu-wpcom' ),
 			),
 		)
 	);
 }
-add_action( 'admin_bar_menu', 'wpcom_repurpose_wp_logo_as_all_sites_menu', 11 );
+add_action( 'admin_bar_menu', 'wpcom_replace_wp_logo_with_wpcom_all_sites_menu', 11 );
 
 /**
  * Adds the Reader menu.
@@ -86,7 +109,7 @@ function wpcom_add_reader_menu( $wp_admin_bar ) {
 		array(
 			'id'    => 'reader',
 			'title' => __( 'Reader', 'jetpack-mu-wpcom' ),
-			'href'  => 'https://wordpress.com/read',
+			'href'  => maybe_add_origin_site_id_to_url( 'https://wordpress.com/read' ),
 			'meta'  => array(
 				'class' => 'wp-admin-bar-reader',
 			),
@@ -94,6 +117,27 @@ function wpcom_add_reader_menu( $wp_admin_bar ) {
 	);
 }
 add_action( 'admin_bar_menu', 'wpcom_add_reader_menu', 15 );
+
+/**
+ * Points the (Profile) -> Edit Profile menu to /me when appropriate.
+ *
+ * @param WP_Admin_Bar $wp_admin_bar The WP_Admin_Bar core object.
+ */
+function wpcom_maybe_replace_edit_profile_menu_to_me( $wp_admin_bar ) {
+	$edit_profile_node = $wp_admin_bar->get_node( 'user-info' );
+	if ( $edit_profile_node ) {
+		// If one of the following is true:
+		// - the user is not a member of the current site
+		// - the current site uses Default admin interface
+		//
+		// Then, the Edit Profile menu should point to /me, instead of the site's profile.php.
+		if ( ! is_user_member_of_blog() || get_option( 'wpcom_admin_interface' ) !== 'wp-admin' ) {
+			$edit_profile_node->href = maybe_add_origin_site_id_to_url( 'https://wordpress.com/me' );
+			$wp_admin_bar->add_node( (array) $edit_profile_node );
+		}
+	}
+}
+add_action( 'admin_bar_menu', 'wpcom_maybe_replace_edit_profile_menu_to_me', 1 );
 
 /**
  * Adds (Profile) -> My Account menu pointing to /me/account.
@@ -106,13 +150,12 @@ function wpcom_add_my_account_item_to_profile_menu( $wp_admin_bar ) {
 		// Adds the 'My Account' menu item before 'Log Out'.
 		$wp_admin_bar->remove_node( 'logout' );
 	}
-
 	$wp_admin_bar->add_node(
 		array(
 			'id'     => 'wpcom-profile',
 			'parent' => 'user-actions',
 			'title'  => __( 'My Account', 'jetpack-mu-wpcom' ),
-			'href'   => 'https://wordpress.com/me/account',
+			'href'   => maybe_add_origin_site_id_to_url( 'https://wordpress.com/me/account' ),
 		)
 	);
 
@@ -121,3 +164,19 @@ function wpcom_add_my_account_item_to_profile_menu( $wp_admin_bar ) {
 	}
 }
 add_action( 'admin_bar_menu', 'wpcom_add_my_account_item_to_profile_menu' );
+
+/**
+ * Replaces the default admin bar class with our own.
+ *
+ * @param string $wp_admin_bar_class Admin bar class to use. Default 'WP_Admin_Bar'.
+ * @return string Name of the admin bar class.
+ */
+function wpcom_custom_wpcom_admin_bar_class( $wp_admin_bar_class ) {
+	if ( get_option( 'wpcom_admin_interface' ) === 'wp-admin' ) {
+		return $wp_admin_bar_class;
+	}
+
+	require_once __DIR__ . '/class-wpcom-admin-bar.php';
+	return '\Automattic\Jetpack\Jetpack_Mu_Wpcom\WPCOM_Admin_Bar';
+}
+add_filter( 'wp_admin_bar_class', 'wpcom_custom_wpcom_admin_bar_class' );