+++ projects/packages/jetpack-mu-wpcom/src/features/wpcom-sidebar-notice/wpcom-sidebar-notice.js
@@ -1,16 +1,18 @@
 /* global wp, wpcomSidebarNotice */
+import { wpcomTrackEvent } from '../../common/tracks';
+
+import './wpcom-sidebar-notice.scss';
 
 const wpcomSidebarNoticeRecordEvent = event => {
 	if ( ! event ) {
 		return;
 	}
-	window._tkq = window._tkq || [];
-	window._tkq.push( [
-		'identifyUser',
+	wpcomTrackEvent(
+		event.name,
+		event.props,
 		wpcomSidebarNotice.user.ID,
-		wpcomSidebarNotice.user.username,
-	] );
-	window._tkq.push( [ 'recordEvent', event.name, event.props ] );
+		wpcomSidebarNotice.user.username
+	);
 };
 
 const wpcomShowSidebarNotice = () => {