+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/breve/store/actions.ts
@@ -3,7 +3,8 @@
  */
 import { askQuestionSync } from '@automattic/jetpack-ai-client';
 import { select } from '@wordpress/data';
-import { getRequestMessages } from '../utils/getRequestMessages';
+import { BREVE_FEATURE_NAME } from '../constants';
+import { getRequestMessages } from '../utils/get-request-messages';
 
 // ACTIONS
 
@@ -48,20 +49,43 @@ export function toggleFeature( feature: string, force?: boolean ) {
 	};
 }
 
+export function setBlockMd5( feature: string, blockId: string, md5: string ) {
+	return {
+		type: 'SET_BLOCK_MD5',
+		feature,
+		blockId,
+		md5,
+	};
+}
+
+export function invalidateSuggestions( feature: string, blockId: string ) {
+	return {
+		type: 'INVALIDATE_SUGGESTIONS',
+		feature,
+		blockId,
+	};
+}
+
 export function setSuggestions( {
+	anchor,
 	id,
 	feature,
 	target,
-	sentence,
+	text,
 	blockId,
+	occurrence,
 }: {
+	anchor: HTMLElement;
 	id: string;
 	feature: string;
 	target: string;
-	sentence: string;
+	text: string;
 	blockId: string;
+	occurrence: string;
 } ) {
 	return ( { dispatch } ) => {
+		anchor?.classList?.add( 'is-loading' );
+
 		dispatch( {
 			type: 'SET_SUGGESTIONS_LOADING',
 			id,
@@ -74,14 +98,17 @@ export function setSuggestions( {
 			getRequestMessages( {
 				feature,
 				target,
-				sentence,
+				text,
 				blockId,
+				occurrence,
 			} ),
 			{
-				feature: 'jetpack-ai-breve',
+				feature: BREVE_FEATURE_NAME,
 			}
 		)
 			.then( response => {
+				anchor?.classList?.remove( 'is-loading' );
+
 				try {
 					const suggestions = JSON.parse( response );
 					dispatch( {
@@ -102,6 +129,8 @@ export function setSuggestions( {
 				}
 			} )
 			.catch( () => {
+				anchor?.classList?.remove( 'is-loading' );
+
 				dispatch( {
 					type: 'SET_SUGGESTIONS_LOADING',
 					id,